// src/app/BusinessModel/Tennis/tennis-score-info-type-enum.enum.ts
export enum TennisScoreInfoTypeEnum {
    CS = 9,
    CFS = 10,
    Set1 = 11,
    Set2 = 12,
    Set3 = 13,
    Set4 = 14,
    Set5 = 15,
    CP = 16,
    Finished = 17,
    FinishedOverTime = 23,
    FinishedAfterPenalty = 24,
    FirstPeriod = 25,
    SecondPeriod = 26,
    ThirdPeriod = 27,
    ExtraTime = 28
}